move things around to stay in existing paradigm
